Butterfly/butterflyApp/manager/requirements.txt
Flask==1.0.3
kafka-python==1.4.6
mongoengine==0.17.0